changing 'count'/size to NSUInteger to conform with NSArray, adding array conversion function

+++ Classes/CPCollection.h
@@ -10,6 +10,6 @@
 @protocol CPCollection <NSFastEnumeration>
 
 -(BOOL)isEmpty;
--(int)count;
+-(NSUInteger)count;
 
 @end
